[TASK] Add 11.5 branch and restore core main (v12) support
This change adds the branch aliasing for v11 (11 => 11.5)
and additional tells the extractor that we have a "v12"
version, which is picked up for "main" branch. This should
restore the ability to get translation files in current
v12 monorepo instances and make backend language selectable
for users, which is not possible with missing language packs.

Resolves #4