Clean up TranslationApi class
`getProjectBuilds()` called in method `getLastFinishedBuildId()`
returns a `ModelCollection` object. Therefore, the check for
`!$items` was always false. As `ModelCollection` implements
`Countable` we can fix the condition with a `count()` call.

The docblock annotations for `getLastFinishedBuildId()` are redundant
and have been removed.

Additionally, the method `getBuilds()` was not used in the project
and is dropped.