Delegate logic of `app:build` command into dedicated service class
The `app:build` command has been refactored to be more controller-like: it should
not execute the logic itself, but call a dedicated service (`Builder`) for that.

Now the streamlined logic of the command can also be tested. As these are
the first tests, phpunit has been introduced as dev dependency.

To detangle the progress bar logic from the service class, a DTO `Progress`
has been created which provides closures, that are called at the appropriate
places to start, advance and finish the progress.

We still have the `SymfonyStyle $io` variable in the `Builder` class as a
dependency to the command. This can be refactored in a separate patch.

Additionally:
- The `PagesNotFoundException` class and the `Progress` class have been
  placed in the `Build` folder for the time being. They may move later to
  another place with more refactorings that also have a need for these classes.
- The `hidden=false` attribute parameter has been removed as this is the default.
- A new Composer command `tests` has been introduced which runs the tests.
- A new Composer command `all` has been introduced to run all quality checks at once.
- A new task `Run tests` has been added to the CI GitHub action to ensure
  the all tests pass on changes.