Clean up CoreInformation class
This class is used by the extract core command.

Some methods are not in use anymore and has been dropped:
- `getAllVersions()`
- `getBranchNameForVersion()`
- `getCoreExtensionKeys()`

The iterable return types have been added, where appropriate.

The `array_search()` function returns `false` on error, not `null`. This has been corrected in the `getVersionForBranchName()`.

For better readability each extension key in the `CORE_EXTENSIONS` constant is now in a separate line.

As the class holds no state it is marked as readonly.