Use crowdin-base library
The Crowdin base library is used. Therefore, the `app:setup` command is superfluous and removed in favor of the `bin/crowdinSetup` console command.

Now, we have a `\App\Configuration\ProjectCollection` class which holds the `\FriendsOfTYPO3\CrowdinBase\Configuration\Entity\Project` configuration of each
Crowdin project. To be able to inject the `ProjectCollection` a `ProjectCollectionFactory` is used to build the collection via Symfony DI.

As first command, the `app:build` command has be adapted to use the new `ProjectCollection` class.