Reorganize workflows
The previous crowdin-build workflow was superfluous: it runs the setup to create the configuration.json. The build command was deactivated. So the workflow had no value.

Running the build command has been moved to the extract workflow: Previously (with v2) at minute 0 the projects were built. Then at minute 30 the projects were extracted. But if translations were changed between those two runs, they were picked up in the next run two hours later. Therefore, the build command is not executed in the same workflow just before extracting the projects.

The extract workflow runs now on minute 15, as GitHub actions documentation states that many scripts run at minute 0 and there may be a delay on executing due to high load.

Additionally, the "crowdin" prefix in the workflow names are superfluous and therefore removed.