Avoid pollution of Builder class with command-specific classes (#49)
Previously, to the Builder class the verbose flag and the SymfonyStyle ($io) has been passed. As the Build class should be agnostic about command-specific classes, a dedicated output class is passed which takes care of the output handling.

Additionally, with these output classes we can reuse the logic for the other commands later and have therefore a similar output for every command.