Introduce TranslationRepository to connect to Crowdin's Translation API
A `TranslationRepository` is introduced to connect to the Translation API. The `ExportService` has become obsolete and also the `TranslationApi::buildProject()` method. This avoids one more indirection. Additionally, the Crowdin client no gets the access token injected automatically - we can just inject the client in a class and can start working.

Additionally, the success message of the `app:build` command has been adjusted as there might be errors (which are output at the end) and the current message does not reflect this.