Use correct language structure in ProjectConfiguration
With the change of providing languages as an array (id, name) in the configuration.json,
the `ProjectConfiguration` was overlooked. Now, the languages are mapped to a list of string
containing the id.

Related: https://github.com/TYPO3/crowdin-bridge/commit/081f5ecc747d78d2ca698aa4b0457ce9eb63fd3c