Pass optional project identifier to project status command
Instead passing the extension key, now the project identifier has to be passed. This is then in line with other commands which expect the project identifier. This way, the project object can be retrieved via the identifier.

Additionally, the project object if passed to the export method to avoid an additional lookup for the project.